refactor: model 구조 변경
