+++ app/src/main/java/com/teamopendata/mindcareapp/model/entity/Record.java
@@ -1,4 +1,4 @@
-package com.teamopendata.mindcareapp.model;
+package com.teamopendata.mindcareapp.model.entity;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
